% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csSAM.R
\name{csSAM}
\alias{csSAM}
\title{csSAM}
\usage{
csSAM(ghat1, se1, n1, ghat2, se2, n2, standardize, medianCenter = TRUE,
  nonNeg = FALSE, stat.only = TRUE)
}
\arguments{
\item{ghat1}{Expression matrix of deconvolved cell-specific gene expression
estimates for group 1.}

\item{se1}{Standard error group 1}

\item{n1}{Group 1 size}

\item{ghat2}{Expression matrix of deconvolved cell-specific gene expression
estimates for group 2.}

\item{se2}{Standard error group 2}

\item{n2}{Group 2 size}

\item{standardize}{Standardize contrast values, into SAM shrinked statistic.}

\item{medianCenter}{Median center rhat distributions for each cell-type}

\item{nonNeg}{Negative values not allowed such as in a single channel
microarray. Zero them if negative (a conervative option)}

\item{stat.only}{logical that indicates if the returned object should consists in the
matrix of cell type-specific SAM statistic only, or contain more detailed data about
the fit (see section \emph{Value}).}
}
\value{
A matrix object that contains the -- standardised -- contrast of the average
cell-specific expression profile of the two groups, per cell-type (Size k by
g where k is the number of cells and g is the number of genes).

If \code{stat.only=FALSE}, then result is a list with the following elements, where
matrix elements have cell types in rows and genes in columns:
\describe{
\item{coefficients}{a matrix of cell-specific contrasts between the two groups}
\item{se}{a matrix of cell-specific standard errors}
\item{s0}{a matrix of cell-specific standard error SAM shrinkage}
\item{stat}{a matrix of -- possibly median-centered -- cell-specific SAM shrinked t-test statistics}
\item{stat.med}{a vector containing the medians of each cell-specific SAM shrinked t-test statistics}
}
}
\description{
Computes the constrast between groups for the deconvolved cell-specific
expression for each cell-type
}
\references{
Shen-Orr SS, Tibshirani R, Khatri P, Bodian DL, Staedtler F, Perry NM, Hastie T, Sarwal MM, Davis MM and Butte AJ (2010).
“Cell type-specific gene expression differences in complex tissues.” \emph{Nature methods}, \emph{7}(4), pp. 287-9. ISSN 1548-7105,
doi: 10.1038/nmeth.1439 (URL: http://doi.org/10.1038/nmeth.1439).
}
\author{
Shai Shen-Orr, Rob Tibshirani, Narasimhan Balasubramanian, David Wang
}
