% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csSAMfit.R
\name{csSAMfit}
\alias{csSAMfit}
\alias{csSAMfit.formula}
\alias{csSAMfit.default}
\title{Fits a csSAM Model}
\usage{
csSAMfit(x, ...)

\method{csSAMfit}{formula}(x, data = NULL, ...)

\method{csSAMfit}{default}(x, cc, y = NULL, covariates = NULL,
  nperms = 200, alternative = c("all", "two.sided", "greater", "less"),
  fit = c("auto", "block", "lm", "monovariate"), standardize = TRUE,
  medianCenter = TRUE, logRm = FALSE, logBase = 2, nonNeg = NULL,
  keep.all = FALSE, verbose = FALSE, .rng = NULL, ...)
}
\arguments{
\item{x}{target matrix assayed on mixed samples.}

\item{...}{arguments passed down to the relevant S3 method.}

\item{cc}{Matrix of cell-frequency. (n by k, n samples, k cell-types)}

\item{y}{A numeric vector of group association of each sample. Either 1 or 2.}

\item{covariates}{matrix of covariates to include in the model, e.g., for correcting
for gender, age or batch effects.}

\item{nperms}{The number of permutations to perform.}

\item{alternative}{two.sided less greater}

\item{fit}{fitting method, that dictates the type of model that is fitted:
\itemize{
\item auto: automatic selection based on input data and model specification (e.g., presence of covariates)
\item block: two linear models are fitted within each group of samples. Only works when the variable
of interest is a factor (categorical variable).
\item lm: full interaction linear model.
\item monovariate: model is fitted including one cell type at a time.
This could be a work around for when there are many more cell types than samples in a given group.
}}

\item{standardize}{Standardize sample or not. Default is TRUE.}

\item{medianCenter}{Median center rhat distributions. Default is TRUE.}

\item{logRm}{Exponentiate data for deconvolution stage. Default is FALSE}

\item{logBase}{Base of logaritm used to determine exponentiation factor.
Default is 2}

\item{nonNeg}{For single channel arrays. Set any cell-specific expression
estimated as negative, to a ceiling of 0. It is conservative in its study of
differential expression. Default is TRUE.}

\item{keep.all}{logical that indicates if the statistics of all the permutations should be also
returned in slot \code{rhatperm}.}

\item{verbose}{single logical or integer that controls verbosity levels.}

\item{.rng}{specification of the initial RNG settings, that are set using \link[rngtools:setRNG]{rngtools::setRNG}.}
}
\value{
a \code{csSAMfit} object, that is a list with the following elements:
*
*
*
}
\description{
Fits a csSAM Model
}
\section{Methods (by class)}{
\itemize{
\item \code{formula}: convenient formula interface that can take advantage of expression and
sample annotation data stored in \link[Biobase:ExpressionSet]{Biobase::ExpressionSet} objects.

\item \code{default}: workhorse method that takes model elements in separate
arguments.
}}

