% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csstats.R
\name{csTopTable}
\alias{csTopTable}
\alias{csTopTable.default}
\alias{csTopTable.array}
\alias{csTopTable.character}
\alias{csTopTable.csSAMfit}
\title{Compute Cell-Specific Statistics}
\usage{
csTopTable(x, ...)

\method{csTopTable}{default}(x, n = 100L, sort.by = 1L,
  decreasing = FALSE, select = NULL, subset = names(threshold),
  threshold = NULL, annotation = NULL, ..., merge = FALSE,
  nodups = FALSE)

\method{csTopTable}{array}(x, ...)

\method{csTopTable}{character}(x, ...)

\method{csTopTable}{csSAMfit}(x, alternative = c("two.sided", "greater",
  "less", "min"), sort.by = "FDR", decreasing = FALSE, ...)
}
\arguments{
\item{x}{data object, typically obtained by fitting a model that computes
cell type-specific statistics, e.g., \link{csSAMfit}.}

\item{...}{extra parameters passed to specific methods.}

\item{n}{maximum number of top features to retain -- in each cell type.
If \code{NULL} or \code{Inf} then all selected features are returned.
Otherwise the first \code{n} feature in each cell type are returned, \strong{after}
ordering according to arguments \code{sort.by} and \code{decreasing}.}

\item{sort.by}{variable to use to order the features.}

\item{decreasing}{logical that controls the feature ordering based on \code{sort.by}.}

\item{select}{specifies the statistic to include in the result.}

\item{subset}{indicates the subset of tissue/cell type to include in the result.
Default is to return only the cell types for which a \code{threshold} value was
provided. If \code{NULL}, as default, then no subsetting is performed.}

\item{threshold}{numeric vector that indicates the threshold applied to select
features based on their values of \code{sort.by}.
Thresholds can be defined for specific cell types by using naming the elements
of \code{threshold} with corresponding cell type names.
The default threshold that is applied to all cell types is taken from the first
unnamed element.

Depending on \code{decreasing}, thresholds are interpreted differently:
\itemize{
\item \code{decreasing = FALSE} keeps features with \code{sort.by} "<= x"
\item \code{decreasing = TRUE} keeps features with \code{sort.by} ">= x"
}}

\item{annotation}{Bioconductor annotation package to use to annotate
the features (rownames) using \link[xbioc:geneInfo]{xbioc::geneInfo}.}

\item{merge}{logical that indicates if the result should be a
a list of \code{data.frame} objects (\code{merge = FALSE}), or a single
\code{data.frame} object with columns "Feature" and "Cell type" containing
the feature identifiers and cell type respectively.}

\item{nodups}{a logical that indicates if features appearing in multiple
cell types should be removed.
This is useful in combination with \code{threshold} to extract features that are
found significant in a single cell type.}

\item{alternative}{Alternative hypothesis to consider.
Alternative "min" returns, for each features, the alternative achieving the minimum
FDR.}
}
\value{
a list of \code{data.frame} objects if \code{merge = FALSE}, or a single
\code{data.frame} object if \code{merge = TRUE}.

If \code{merge=TRUE} and \code{select} is of length one, then the result is a \code{matrix} object
(feature x cell type) containing the selected value.
}
\description{
\code{csTopTable} is a generic function that returns cell type-level statistics
on biological features, e.g., cell type-specific expression.
}
\section{Methods (by class)}{
\itemize{
\item \code{csSAMfit}: returns, for each feature, the false discovery
rates of differential expression between groups of samples within each cell type,
as computed by \code{\link[=csSAM]{fdrCsSAM}} when running csSAM.
These are returned as a list, whith one element per cell type.
}}

\seealso{
\link{fdrCsSAM}
}
