% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findSigGene.R
\name{findSigGene}
\alias{findSigGene}
\title{findSigGene}
\usage{
findSigGene(rhat, csSAMData, alternative = c("two.sided", "less", "greater"),
  version = getOption("csSAM.siggene.version", 2))
}
\arguments{
\item{rhat}{Matrix of cell-specific contrasts for each gene in each cell-type
as computed for the original group classification.}

\item{csSAMData}{List object returned from fdrCsSAM.}

\item{alternative}{alternative hypothesis being tested: \code{'two.sided'}, \code{'less'}
or \code{'greater'}}

\item{version}{number that indicates the version of computation to perform: 1 or 2.
Current default is \code{version = 2}, which fixes an issue for alternatives
\code{'less'} and \code{'greater'}. Use \code{version = 1} to obtain old behaviour.}
}
\value{
A matrix size k by g where k is the number of cell-types and g is the
number of genes. For each cell in the matirx, listed is the FDR of the gene
for a difference in a given cell-type.
}
\description{
Find the false discovery rate for each gene in each cell-type.
}
\note{
From version 1.3, the FDR computation is implemented in \emph{C++},
using \pkg{Rcpp}.
}
\references{
Shen-Orr SS, Tibshirani R, Khatri P, Bodian DL, Staedtler F, Perry NM, Hastie T, Sarwal MM, Davis MM and Butte AJ (2010).
“Cell type-specific gene expression differences in complex tissues.” \emph{Nature methods}, \emph{7}(4), pp. 287-9. ISSN 1548-7105,
doi: 10.1038/nmeth.1439 (URL: http://doi.org/10.1038/nmeth.1439).
}
\author{
Shai Shen-Orr, Rob Tibshirani, Narasimhan Balasubramanian, David Wang

\emph{C++} implementation/optimisation by Renaud Gaujoux
}
