% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCsSAM.R
\name{plot.csSAMfit}
\alias{plot.csSAMfit}
\title{Plotting csSAM FDR}
\usage{
\method{plot}{csSAMfit}(x, subset = NULL, alternative = "two.sided",
  by = c("cell-type", "alternative"), xlab = "# called", ylab = "FDR",
  ylim = c(0, 1), ...)
}
\arguments{
\item{x}{a csSAM fit as returned by \link{csSAMfit}}

\item{subset}{character or index vector that specifies the subset of
cell types to include in the plot.}

\item{alternative}{test alternative hypothesis for which FDRs are shown.
Multiple alternatives can be plotted at once ("two.sided", "greater", "less").
Use \code{alternative = "all"} to plot all of them.}

\item{by}{specifies the facetting criteria.
Default is produces one panel per cell type, each showing FDR for the selected alternative(s).}

\item{xlab, ylab}{axis labels}

\item{ylim}{y-axis range}

\item{...}{unused trailing arguments for S3 conformity.
An error will be thrown if any unsused argument are passed to the function.}
}
\description{
Plotting csSAM FDR
}
