% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCsSAM.R
\name{plotCsSAM}
\alias{plotCsSAM}
\title{plotcsSAM}
\usage{
plotCsSAM(csfit, sam, alternative = "two.sided", fileName, ..., height = 8,
  width = 10)
}
\arguments{
\item{csfit}{csSAM fitted model as returned by \link{csSAMfit}.}

\item{sam}{List object output of the fdrSAM function}

\item{alternative}{Type of test conducted. Will appear in plot title.}

\item{fileName}{Name of output pdf file.}

\item{...}{other arguments passed to \link{plot.csSAMfit}.}

\item{height, width}{output height and width}
}
\description{
Plots the # of genes called significnat at a given false disocvery rate for
the SAM (heterogenous tissue) comparison, and for each of the contrasted
cell-types using csSAM
}
\references{
Shen-Orr SS, Tibshirani R, Khatri P, Bodian DL, Staedtler F, Perry NM, Hastie T, Sarwal MM, Davis MM and Butte AJ (2010).
“Cell type-specific gene expression differences in complex tissues.” \emph{Nature methods}, \emph{7}(4), pp. 287-9. ISSN 1548-7105,
doi: 10.1038/nmeth.1439 (URL: http://doi.org/10.1038/nmeth.1439).
}
\author{
Shai Shen-Orr, Rob Tibshirani, Narasimhan Balasubramanian, David Wang
}
