\name{csSAM-internal}
\alias{csSAM-internal}
\alias{make.monotone}
\alias{ttest.func}
\alias{varr}
\title{Internal csSAM functions}
\usage{
  make.monotone(x)

  ttest.func(x, y, s0 = 0, sd = NULL)

  varr(x, meanx = NULL)
}
\description{
  These functions are not to be called by the user.
}
\author{
  Shai Shen-Orr, Rob Tibshirani, Narasimhan
  Balasubramanian, David Wang
}
\keyword{internal}

